refactor in process
