+++ stress/common/src/main/java/alluxio/stress/rpc/RpcBenchParameters.java
@@ -1,0 +1,14 @@
+package alluxio.stress.rpc;
+
+import alluxio.stress.Parameters;
+import com.beust.jcommander.Parameter;
+
+public class RpcBenchParameters extends Parameters {
+  @Parameter(names = {"--concurrency"},
+          description = "simulate this many workers on one machine")
+  public int mConcurrency = 2;
+
+  @Parameter(names = {"--duration"},
+          description = "The length of time to run the benchmark. (1m, 10m, 60s, 10000ms, etc.)")
+  public String mDuration = "30s";
+}
